feat: update search bar to fc
